# docs/openapi.yaml
openapi: 3.0.3
info:
  title: gosso - SSO Authentication Server
  description: |
    A production-ready Single Sign-On (SSO) authentication server built with Go.

    Supports:
    - Username/password login with MFA (TOTP + backup codes)
    - OAuth 2.0 Authorization Server (authorization code, refresh token, client credentials)
    - OpenID Connect Provider (ID token, discovery, JWKS, userinfo)
    - Social login (Google, GitHub, WeChat)
    - Admin account management
    - Email/phone verification and password reset
  version: 1.0.0
  contact:
    name: rushairer
    url: https://github.com/rushairer/gosso
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchanges a valid refresh token for a new access token and refresh token pair.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /oauth2/authorize:
    get:
      tags: [OAuth2 Protocol]
      summary: OAuth2 Authorization endpoint
      description: |
        Initiates the OAuth2 authorization code flow.
        If the user has already consented, redirects with the authorization code.
        Otherwise, renders the consent form (HTML).
        Requires authentication (JWT).
      operationId: oauth2Authorize
      security:
        - BearerAuth: []
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: scope
          in: query
          schema:
            type: string
          description: Space-separated list of scopes
        - name: state
          in: query
          schema:
            type: string
          description: Opaque value for CSRF protection
        - name: code_challenge
          in: query
          schema:
            type: string
          description: PKCE code challenge
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256, plain]
          description: PKCE code challenge method
        - name: nonce
          in: query
          schema:
            type: string
          description: OIDC nonce
      responses:
        "302":
          description: Redirect to redirect_uri with authorization code
          headers:
            Location:
              schema:
                type: string
                example: https://app.example.com/callback?code=abc123&state=xyz
        "200":
          description: Consent form (HTML)
          content:
            text/html:
              schema:
                type: string
        "400":
          description: OAuth2 error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2Error"

    post:
      tags: [OAuth2 Protocol]
      summary: Submit consent decision
      description: Submits the user's consent decision for the OAuth2 authorization flow.
      operationId: oauth2SubmitConsent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [client_id, redirect_uri, scope, approved]
              properties:
                client_id:
                  type: string
                redirect_uri:
                  type: string
                scope:
                  type: string
                state:
                  type: string
                approved:
                  type: string
                  enum: ["true", "false"]
                code_challenge:
                  type: string
                code_challenge_method:
                  type: string
                nonce:
                  type: string
      responses:
        "302":
          description: Redirect to redirect_uri with authorization code
        "400":
          description: OAuth2 error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2Error"
  /oauth2/token:
    post:
      tags: [OAuth2 Protocol]
      summary: OAuth2 Token endpoint
      description: |
        Exchanges an authorization code, refresh token, or client credentials for tokens.
        Supports grant types: `authorization_code`, `refresh_token`, `client_credentials`.
        Client authentication via request body or HTTP Basic Auth.
      operationId: oauth2Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Token response (RFC 6749)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: OAuth2 error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2Error"
        "401":
          description: Client authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2Error"
  /oauth2/revoke:
    post:
      tags: [OAuth2 Protocol]
      summary: Revoke a token
      description: Revokes a refresh token. Requires authentication.
      operationId: oauth2Revoke
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: The token to revoke
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /oauth2/introspect:
    post:
      tags: [OAuth2 Protocol]
      summary: Token introspection (RFC 7662)
      description: |
        Introspects an access or refresh token. Client authentication via HTTP Basic Auth or form parameters.
        Returns token metadata if active, or `{"active": false}` if not.
      operationId: oauth2Introspect
      security:
        - BasicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: The token to introspect
      responses:
        "200":
          description: Introspection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntrospectionResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2Error"
        "401":
          description: Client authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2Error"

  # ──────────────────────────────────────────────
  # OIDC
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login, token, or social login endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Auth with client_id as username and client_secret as password
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection via double-submit cookie pattern.
        1. GET any endpoint to receive `csrf_token` cookie
        2. Read cookie value and send as `X-CSRF-Token` header on POST/PUT/DELETE
        3. Not required when using `Authorization: Bearer` header (JWT auth)

  # ── Parameters ──
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            message: invalid request body
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            message: unauthorized
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: error message
    OAuth2Error:
      type: object
      description: RFC 6749 error response
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - consent_denied
            - server_error
        error_description:
          type: string

    # -- Auth Request Bodies --
...
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    Response_TokenPair:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              example: 900
...
    TokenRequest:
      type: object
      required: [grant_type]
      properties:
        grant_type:
          type: string
          enum: [authorization_code, refresh_token, client_credentials]
        code:
          type: string
          description: Authorization code (for authorization_code grant)
        redirect_uri:
          type: string
          format: uri
          description: Must match the redirect_uri from the authorization request
        client_id:
          type: string
        client_secret:
          type: string
          description: Client secret (for confidential clients)
        code_verifier:
          type: string
          description: PKCE code verifier
        refresh_token:
          type: string
          description: Refresh token (for refresh_token grant)
        scope:
          type: string
          description: Requested scopes (for client_credentials grant)
    TokenResponse:
      type: object
      description: RFC 6749 token response
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900
        scope:
          type: string
          description: Space-separated list of granted scopes
        id_token:
          type: string
          description: OpenID Connect ID token (only when openid scope is requested)
    IntrospectionResponse:
      type: object
      description: RFC 7662 introspection response
      properties:
        active:
          type: boolean
          description: Whether the token is active
        scope:
          type: string
        client_id:
          type: string
        username:
          type: string
        token_type:
          type: string
        exp:
          type: integer
          description: Expiration time (Unix timestamp)
        iat:
          type: integer
          description: Issued-at time (Unix timestamp)
        sub:
          type: string
          description: Subject (account ID)
        aud:
          type: string
          description: Audience (client ID)
        iss:
          type: string
          description: Issuer
        jti:
          type: string
          description: JWT ID

    # -- OIDC --
...